Set BuildLevel() to use Unity Resources
If we want to load a level from xml, we can't use a loose file and have
it work after Building into an exe or app or w/e. As long as the xml
file is in the project's Resources folder, passing the name (sans
extension) into BuildLevel should successfully load the level after
creating a runnable application.